[11.x] Move Defer classes to Concurrency component
Also avoid using global `defer()` function when it can be defined in
environment such as Swoole
